Condensed upgrades to one file
