+++ src/senaite/samplepointlocations/profiles/default/metadata.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <metadata>
-  <version>1.0.4</version>
+  <version>1.0.1</version>
   <dependencies>
     <dependency>profile-senaite.core:default</dependency>
   </dependencies>
