+++ src/senaite/samplepointlocations/upgrade/configure.zcml
@@ -10,29 +10,5 @@
      destination="1.0.1"
      handler="senaite.samplepointlocations.upgrade.v01_00_001.upgrade"
      profile="senaite.samplepointlocations:default"/>
-
- <genericsetup:upgradeStep
-     title="Upgrade to SENAITE SAMPLELOCATIONS 1.0.2"
-     description=""
-     source="1.0.1"
-     destination="1.0.2"
-     handler="senaite.samplepointlocations.upgrade.v01_00_002.upgrade"
-     profile="senaite.samplepointlocations:default"/>
-
- <genericsetup:upgradeStep
-     title="Upgrade to SENAITE SAMPLELOCATIONS 1.0.3"
-     description=""
-     source="1.0.2"
-     destination="1.0.3"
-     handler="senaite.samplepointlocations.upgrade.v01_00_003.upgrade"
-     profile="senaite.samplepointlocations:default"/>
-
- <genericsetup:upgradeStep
-     title="Upgrade to SENAITE SAMPLELOCATIONS 1.0.4"
-     description=""
-     source="1.0.3"
-     destination="1.0.4"
-     handler="senaite.samplepointlocations.upgrade.v01_00_004.upgrade"
-     profile="senaite.samplepointlocations:default"/>
 </configure>
